Adding program files
Copying program files from personal repository to shared BaylorICTHUS repository so everyone has access to this code development/testing program. Although initially developed for VisualStudio, subsequent development of this framework program has been performed within the CodeBlocks IDE since this provides access to multiple compilers instead of the VC++ compiler of Visual Studio that often results in incompatibility issues and unforeseen errors when the associated code is deployed on what is typically a Linux system using a gcc compiler. The current project is configured to use the gcc compiler to avoid such issues.